refactor(export): split three oversized STEP export modules (#3184)
Pure refactor, no behaviour change — the full @ifc-lite/export suite
(63 files / 860 tests, 2/30 skipped) is identical before and after.

- step-property-sets.ts (1061 lines) -> step-property-set-readers.ts
  (PropertySetContext, the per-entity STEP-text readers and owner-history
  resolution), step-property-set-index.ts (the IfcRelDefinesByProperties
  reverse index), step-property-set-generators.ts (the two record
  generators), step-property-set-collection.ts (the overlay-edit
  collection phase), and step-property-sets.ts itself, kept as the
  write/rewrite phase every other module's comments already cite by
  that name.
- step-attribute-mutations.ts (415 lines) -> kept as the source-buffer
  mutation pipeline, with step-attribute-serializers.ts (the two
  per-slot serialize helpers shared with the overlay-created path) and
  step-overlay-attribute-overrides.ts (that overlay-created sibling)
  split out, following the split the module's own doc comment already
  described.
- step-serialization.ts (435 lines) -> property-value-serialization.ts
  (serializePropertyValue, the IFC property-type-name mapping table)
  split out; the rest of the generic STEP token serialization stays.

Every resulting file is under the 400-line house rule, so all three
allowlist rows are removed rather than lowered. `node
scripts/check-module-size.mjs --update` also swept three unrelated
files that had drifted under their recorded budget (headless-backend.ts,
backend-query.ts, component-state.ts) and re-pinned ALLOWLIST_DIGEST in
the same commit.
