+++ scripts/check-module-size.mjs
@@ -135,7 +135,7 @@ const SOURCE_RE = /\.(ts|tsx|mts|cts)$/;
  * failure message. Either way do it at the moment you finalise the change — it
  * moves if anything else touched the allowlist first.
  */
-const ALLOWLIST_DIGEST = '1125203688214438683';
+const ALLOWLIST_DIGEST = '16922803857285994499';
 
 function parseArgs(argv) {
   const out = {
